+++ tools/api-markdown-documenter/src/api-item-transforms/configuration/DocumentationSuiteOptions.ts
@@ -8,7 +8,6 @@ import {
 	type ApiItem,
 	ApiItemKind,
 	type ApiPackage,
-	type Excerpt,
 	ReleaseTag,
 } from "@microsoft/api-extractor-model";
 
@@ -18,6 +17,9 @@ import {
 	getUnscopedPackageName,
 	getSafeFilenameForName,
 	getConciseSignature,
+	getSingleLineExcerptText,
+	isDeprecated,
+	getReleaseTag,
 } from "../../utilities/index.js";
 
 /**
@@ -187,6 +189,17 @@ export interface DocumentationSuiteOptions {
 	getLinkTextForItem?: (apiItem: ApiItem) => string;
 
 	/**
+	 * Generate a list of "alerts" to display in API items tables for a given API item.
+	 *
+	 * @param apiItem - The API item for which table cell contents are being generated.
+	 *
+	 * @returns The list of "alert" strings to display.
+	 *
+	 * @defaultValue {@link DefaultDocumentationSuiteOptions.defaultGetAlertsForItem}
+	 */
+	getAlertsForItem?: (apiItem: ApiItem) => string[];
+
+	/**
 	 * Whether or not the provided `ApiPackage` should be skipped during documentation generation.
 	 *
 	 * @param apiPackage - The package that may or may not be skipped.
@@ -304,7 +317,8 @@ export namespace DefaultDocumentationSuiteOptions {
 				// For signature items, the display-name is not particularly useful information
 				// ("(constructor)", "(call)", etc.).
 				// Instead, we will use a cleaned up variation on the type signature.
-				return getSingleLineExcerptText((apiItem as ApiDeclaredItem).excerpt);
+				const excerpt = getSingleLineExcerptText((apiItem as ApiDeclaredItem).excerpt);
+				return trimTrailingSemicolon(excerpt);
 			}
 			default: {
 				return apiItem.displayName;
@@ -328,12 +342,39 @@ export namespace DefaultDocumentationSuiteOptions {
 				// For signature items, the display-name is not particularly useful information
 				// ("(constructor)", "(call)", etc.).
 				// Instead, we will use a cleaned up variation on the type signature.
-				return getSingleLineExcerptText((apiItem as ApiDeclaredItem).excerpt);
+				const excerpt = getSingleLineExcerptText((apiItem as ApiDeclaredItem).excerpt);
+				return trimTrailingSemicolon(excerpt);
 			}
 			default: {
 				return getConciseSignature(apiItem);
 			}
 		}
+	}
+
+	/**
+	 * Default {@link DocumentationSuiteOptions.getHeadingTextForItem}.
+	 *
+	 * Generates alerts for the following tags, if found:
+	 *
+	 * - `@alpha`: "Alpha"
+	 *
+	 * - `@beta`: "Beta"
+	 *
+	 * - `@deprecated`: "Deprecated"
+	 */
+	export function defaultGetAlertsForItem(apiItem: ApiItem): string[] {
+		const alerts: string[] = [];
+		if (isDeprecated(apiItem)) {
+			alerts.push("Deprecated");
+		}
+
+		const releaseTag = getReleaseTag(apiItem);
+		if (releaseTag === ReleaseTag.Alpha) {
+			alerts.push("Alpha");
+		} else if (releaseTag === ReleaseTag.Beta) {
+			alerts.push("Beta");
+		}
+		return alerts;
 	}
 
 	/**
@@ -358,26 +399,10 @@ const defaultDocumentationSuiteOptions: Required<DocumentationSuiteOptions> = {
 	getUriBaseOverrideForItem: DefaultDocumentationSuiteOptions.defaultGetUriBaseOverrideForItem,
 	getHeadingTextForItem: DefaultDocumentationSuiteOptions.defaultGetHeadingTextForItem,
 	getLinkTextForItem: DefaultDocumentationSuiteOptions.defaultGetLinkTextForItem,
+	getAlertsForItem: DefaultDocumentationSuiteOptions.defaultGetAlertsForItem,
 	skipPackage: DefaultDocumentationSuiteOptions.defaultSkipPackage,
 	minimumReleaseLevel: ReleaseTag.Internal, // Include everything in the input model
 };
-
-/**
- * Extracts the text from the provided excerpt and adjusts it to be on a single line, and to omit any trailing `;`.
- *
- * @privateRemarks If we find that this is useful in more places, we might consider moving this to a
- * public utilities module and make it part of the public helper suite.
- */
-function getSingleLineExcerptText(excerpt: Excerpt): string {
-	// Regex replaces line breaks with spaces to ensure everything ends up on a single line.
-	let signatureExcerpt = excerpt.text.trim().replace(/\s+/g, " ");
-
-	if (signatureExcerpt.endsWith(";")) {
-		signatureExcerpt = signatureExcerpt.slice(0, signatureExcerpt.length - 1);
-	}
-
-	return signatureExcerpt;
-}
 
 /**
  * Gets a complete {@link DocumentationSuiteOptions} using the provided partial configuration, and filling
@@ -390,4 +415,14 @@ export function getDocumentationSuiteOptionsWithDefaults(
 		...defaultDocumentationSuiteOptions,
 		...inputOptions,
 	};
+}
+
+/**
+ * Trims a trailing semicolon from the provided text, if the text contains one.
+ */
+function trimTrailingSemicolon(text: string): string {
+	if (text.endsWith(";")) {
+		return text.slice(0, text.length - 1);
+	}
+	return text;
 }